requirements.txt
#asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-createsuperuser==2020.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
dnspython==2.2.1
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
Pillow==9.2.0
PyJWT==2.4.0
pymongo==4.2.0
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.2.4
uritemplate==4.1.1
urllib3==1.26.13
whitenoise==6.2.0
django-storages>=1.13.2
boto3
boto3>=1.28.0
bcrypt==3.2.0
djangorestframework==3.13.1
